[MODIFY] change printf parameter

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,9 +9,9 @@
 int main(void)
 {
 	int i;
-	unsigned long int ant = 0;
-	unsigned long int new = 0;
-	unsigned long int sum = 1;
+	unsigned long long int ant = 0;
+	unsigned long long int new = 0;
+	unsigned long long int sum = 1;
 
 	for (i = 0; i < 98; i++)
 	{
